Move stats to a separate job
