+++ eth-crawler/src/main/kotlin/org/apache/tuweni/eth/crawler/rest/ClientsService.kt
@@ -19,7 +19,7 @@ package org.apache.tuweni.eth.crawler.rest
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.apache.tuweni.eth.EthJsonModule
 import org.apache.tuweni.eth.crawler.RESTMetrics
-import org.apache.tuweni.eth.crawler.RelationalPeerRepository
+import org.apache.tuweni.eth.crawler.StatsJob
 import javax.servlet.ServletContext
 import javax.ws.rs.GET
 import javax.ws.rs.Path
@@ -44,18 +44,18 @@ class ClientsService {
   @Produces(MediaType.APPLICATION_JSON)
   @Path("all")
   fun getClientIds(): String {
-    val repo = context!!.getAttribute("repo") as RelationalPeerRepository
+    val stats = context!!.getAttribute("stats") as StatsJob
     val metrics = context!!.getAttribute("metrics") as RESTMetrics
     metrics.clientsCounter.add(1)
-    val peers = repo.getClientIds()
+    val peers = stats.getClientIds()
     val result = mapper.writeValueAsString(peers)
     return result
   }
 
   @Path("{upgrade}/stats")
   fun getClientStats(@PathParam("upgrade") upgrade: String): String {
-    val repo = context!!.getAttribute("repo") as RelationalPeerRepository
-    val peers = repo.getUpgradeStats()[upgrade]
+    val stats = context!!.getAttribute("stats") as StatsJob
+    val peers = stats.getUpgradeStats()[upgrade]
     val result = mapper.writeValueAsString(peers)
     return result
   }
@@ -64,8 +64,8 @@ class ClientsService {
   @Produces(MediaType.APPLICATION_JSON)
   @Path("stats")
   fun getClientStats(): String {
-    val repo = context!!.getAttribute("repo") as RelationalPeerRepository
-    val stats = repo.getClientStats()
+    val statsjob = context!!.getAttribute("stats") as StatsJob
+    val stats = statsjob.getClientStats()
     val result = mapper.writeValueAsString(stats)
     return result
   }
